TryOpenexistingworker and createsemaphore moved to semaphore.windows.cs
